Move schema.py and job_executor.py to common/
