+++ tests/agent/test_local_executor.py
@@ -7,7 +7,7 @@ import pytest
 
 from agent import config
 from agent.executors import local, volumes
-from jobrunner.job_executor import ExecutorState, JobDefinition, Privacy, Study
+from common.job_executor import ExecutorState, JobDefinition, Privacy, Study
 from jobrunner.lib import datestr_to_ns_timestamp, docker
 from tests.factories import ensure_docker_images_present, metrics_factory
