+++ tests/agent/test_main.py
@@ -3,7 +3,7 @@ from unittest.mock import ANY, patch
 import pytest
 
 from agent import config, main
-from jobrunner.job_executor import ExecutorState, JobStatus
+from common.job_executor import ExecutorState, JobStatus
 from tests.factories import job_definition_factory, runjob_db_task_factory
