+++ tests/common/test_schema.py
